Added Audio/Visual Service Request and made it accessible by the Service Request landing page.
