+++ .gitignore
@@ -24,6 +24,7 @@ hs_err_pid*
 
 # Gradle #
 .gradle
+.gradle/
 /build/
 
 # Idea #
