+++ src/main/resources/edu/wpi/teamname/Views/AudioVisualServiceRequest.fxml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Text?>
+
+<AnchorPane onMouseEntered="#setup" prefHeight="400.0" prefWidth="599.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.wpi.teamname.views.AudioVisualServiceRequestController">
+   <children>
+      <Text layoutX="222.0" layoutY="27.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Audio/Visual Service Request" />
+      <Text layoutX="24.0" layoutY="58.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Patient Name: " />
+      <Text layoutX="31.0" layoutY="89.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Room:" />
+      <Text layoutX="33.0" layoutY="120.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Floor:" />
+      <TextField fx:id="patientName" layoutX="103.0" layoutY="41.0" />
+      <TextField fx:id="roomNumber" layoutX="74.0" layoutY="72.0" />
+      <TextField fx:id="floor" layoutX="74.0" layoutY="103.0" />
+      <DatePicker fx:id="startDate" layoutX="95.0" layoutY="144.0" />
+      <Text layoutX="29.0" layoutY="160.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Start Date:" />
+      <DatePicker fx:id="endDate" layoutX="95.0" layoutY="183.0" />
+      <Text layoutX="29.0" layoutY="199.0" strokeType="OUTSIDE" strokeWidth="0.0" text="End Date:" />
+      <Text layoutX="29.0" layoutY="241.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Start Time:" />
+      <TextField fx:id="startTime" layoutX="95.0" layoutY="224.0" />
+      <Text layoutX="28.0" layoutY="277.0" strokeType="OUTSIDE" strokeWidth="0.0" text="End Time:" />
+      <TextField fx:id="endTime" layoutX="94.0" layoutY="260.0" />
+      <Button layoutX="265.0" layoutY="349.0" mnemonicParsing="false" onAction="#sendServiceRequest" text="Send" />
+   </children>
+</AnchorPane>
